Move test mocks from utils/ to test/
