+++ test/mocks/database-users/awarding-points.js
