+++ test/mocks/discord/client.js
