+++ test/mocks/discord/index.js
