+++ test/mocks/discord/role.js
