+++ test/mocks/discord/text-channel.js
