+++ test/mocks/discord/user.js
